[core][opengl] Update BufferDesc to use size_t for size and int64_t for stride and offset

+++ include/rndr/core/graphics-types.h
@@ -516,13 +516,13 @@ struct BufferDesc
     Usage usage = Usage::Default;
 
     /** Total size of a buffer in bytes. */
-    uint32_t size = 0;
+    size_t size = 0;
 
     /** Size of one element, in bytes, in an array of elements. */
-    int32_t stride = 0;
+    int64_t stride = 0;
 
     /** Offset, in bytes, from the beginning of the buffer to the first element to use. */
-    uint32_t offset = 0;
+    int64_t offset = 0;
 };
 
 struct SamplerDesc
